fstests: sort-expunges.sh: skip processing symlinks
Skip processing symlinks as they are already processed and if you do
this you also remove the symlink nature.

+++ playbooks/scripts/workflows/fstests/sort-expunges.sh
@@ -25,6 +25,9 @@ FILES=$(find $DIR -name \*.txt)
 
 for i in $FILES; do
 	TMP=${i}.tmp
+	if [ -L $i ]; then
+		continue
+	fi
 	cat $i | sort | uniq > $TMP
 	mv $TMP $i
 done
